AppKeener: Partial notification

+++ AppKeener/Models/Produto.cs
@@ -14,10 +14,11 @@ namespace AppKeener.Models
         [DisplayName("Descrição")]
         [Required(ErrorMessage = "O campo {0} é obrigatório")]
         [StringLength(1000, ErrorMessage = "O campo {0} precisa ter entre {2} e {1} caracteres", MinimumLength = 2)]
-        public string Descricao { get; set; }       
+        public string Descricao { get; set; }
+        
 
         [Required(ErrorMessage = "O campo {0} é obrigatório")]
-        public decimal Valor { get; set; }
+        public float Valor { get; set; }
 
         [Required(ErrorMessage = "O campo {0} é obrigatório")]
         public int Quantidade { get; set; }
